refactor: move config to common folder
