+++ src/modules/telegram/index.ts
@@ -1,6 +1,6 @@
 import { Telegram } from 'telegraf';
 
-import { telegram as telegramConfig } from '../../config';
+import { telegram as telegramConfig } from '../../common/config';
 
 const telegram = new Telegram(telegramConfig.token, {});
